Refactor out LS utilities from XmlNode
