+++ src/common/lewis_and_short/ls.ts
@@ -1,7 +1,7 @@
 import { parse } from "@/common/lewis_and_short/ls_parser";
 import { assert, checkPresent } from "../assert";
 import fs from "fs";
-import { parseEntries, XmlNode } from "./xml_node";
+import { XmlNode } from "./xml_node";
 import { displayEntryFree } from "./ls_display";
 import { getOrths, isRegularOrth, mergeVowelMarkers } from "./ls_orths";
 import { removeDiacritics } from "../text_cleaning";
@@ -10,6 +10,7 @@ import { extractOutline } from "./ls_outline";
 import { Vowels } from "../character_utils";
 import Database from "better-sqlite3";
 import { ServerExtras } from "@/web/utils/rpc/server_rpc";
+import { parseEntries } from "./ls_xml_utils";
 
 interface ProcessedLsEntry {
   keys: string[];
